don't look at newline for calculate bounding box
* rope-read-mode.el(rope-read-y-info-of-line): Stop the iteration before
  the newline character.

The trigger for this change was a dirty reversed display of a line
consisting only of standard characters and the following line starting
with a superscript.  Somehow the superscript geometry was contained in
the newline.

I think this change is a clarification of the function.  The newline
character should not be considered at all for geometry calculations.